STANBOL-783:
Till now Contenthub didn't allow to specify the chain name to be used in the enhancement process of the documents submitted to it. To be able to allow indexing of different enhancements which might be produced by different enhancement engines in different chains, it is necessary to allow configuring the chain name. 

Now, there is an additional "chain" parameter in the SolrStore#enhanceAndPut methods allowing to specify the name of the chain. Based on this, the SolrStoreImpl has been updated. Also, corresponding RESTful services has also been update.

Note that in this commit SolrStoreImpl contains other changes which are basically usage of ContentItem instead of SolrContentItem (sorry for the bulk update).

git-svn-id: https://svn.apache.org/repos/asf/stanbol/trunk@1414034 13f79535-47bb-0310-9956-ffa450edef68